// UTP+Report_copy/front/src/app/api/models/reporte-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ReporteGestionDto } from '../models/reporte-gestion-dto';
export interface ReporteDto {
  contacto?: string;
  descripcion?: string;
  fechaCreacion?: string;
  foto?: string;
  id?: number;
  isAnonimo?: boolean;
  mensajeAdmin?: string;
  mensajeSeguridad?: string;
  reporteGestion?: ReporteGestionDto;
  seguridadAsignadoId?: number;
  tipoIncidenteId?: number;
  usuarioId?: number;
  zonaId?: number;
}