// UTP+Report_copy/front/src/app/api/models/reporte-gestion-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ReporteGestionDto {
  estado?: 'PENDIENTE' | 'EN_PROCESO' | 'UBICANDO' | 'INVESTIGANDO' | 'PENDIENTE_APROBACION' | 'RESUELTO' | 'CANCELADO';
  fechaActualizacion?: string;
  id?: number;
  prioridad?: 'BAJA' | 'MEDIA' | 'ALTA';
}