// UTP+Report_copy/front/src/app/api/models/zona-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ZonaResponse {
  activo?: boolean;
  descripcion?: string;
  estado?: 'ZONA_SEGURA' | 'ZONA_PRECAUCION' | 'ZONA_PELIGROSA';
  foto?: string;
  id?: number;
  nombre?: string;
  reportCount?: number;
  sedeId?: number;
}